Last update

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: ahalleux <ahalleux@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/04 12:39:07 by ahalleux          #+#    #+#             */
-/*   Updated: 2022/04/25 14:22:31 by ahalleux         ###   ########.fr       */
+/*   Updated: 2022/04/25 20:59:28 by ahalleux         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int		i;
+	size_t		i;
 
 	i = 0;
 	while (str[i])
@@ -23,9 +23,3 @@ size_t	ft_strlen(const char *str)
 	}
 	return (i);
 }
-
-// int main()
-// {
-//     char *str = "Hello World";
-//     printf("%i", ft_strlen(str));
-// }
